build.sh: add --clone flag to clone non-existing directories.
If --clone is specified as a parameter to build.sh, non-existing modules
in the build list are automatically cloned into the respective directories.

The current approach requires a user to either use jhbuild to pull all
repositories or to get the git_xorg.sh script from the wiki. By adding clone
support, a user just needs clone the modular module and then run build.sh to
get a development setup.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>