release.sh: remove "git push origin" policy, replace with checks.
Running 'git push origin' at the end of the script will push changes on the
master branch - even if the release is supposed to be a stable branch
release. This may result in not-yet-ready changes being pushed to master.

This patch removes this default git push policy and replaces it with a
warning message if the object belonging to the tag is not on the remote yet.
Requirements: the tag must be on the currently checked-out branch, it is not
possible to cut a branch release from master anymore.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>