Process a user-supplied sub-project list.
New feature.

Using the new "--modlist <file>" option, the user can specify a file which
contains a list of the sub-projects to process. The file can contain blank
lines as well as comment lines which start with a hash mark (#). The
sub-projects will be processed in the order in which they are found in the
list file.

Suggested usage:
    1. run script to generate the list of sub-projects, redirecting to a file
	$ util/modular/build.sh -L > list
    2. edit <list>, commenting-out or deleting unnecessary lines, or adding
       new lines as appropriate
    3. run the build with this list
	$ util/modular/build.sh --modfile list

Reviewed-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>