build.sh: allow tarballs with --modfile
This change allows to build from tarballs when using the new --modfile
user-supplied sub-projects feature. Without this change, using tarballs
was not possible with --modfile

The workflow with the --modfile feature is thus changed to explicitly
require the use of the --clone flag when automatic git clone of the
sub-projects is wanted, which was previously assumed implicitly with
--modfile.

You may shorten the commit message as you see fit.
Cc: xorg-devel, Trevor Woerner, Gaetan Nadon

Signed-off-by: Simon Braunschmidt <sbraun@emlix.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>